ruby_parser can now handle 1.9 syntax for HAML files
I got pretty confused by this error message, as I was getting it with the following code.

```ruby
_("For more information, please contact %{mail}.") { mail: mail_to("test@example.org") }
```

In this case, the problem is not related to 1.9 syntax, but the error message had me thinking it was. ruby_parser now supports 1.9 syntax, so this message is no longer relevant, and it is better to make it more generic.